Extract NaN values and masks per sample
