+++ hist_data_analysis/__init__.py
@@ -1,4 +1,4 @@
 from hist_data_analysis.analytics import analyse,similarity
 from hist_data_analysis.sampling_rate import runme,undersampling_error
-from hist_data_analysis.timeseries.train_and_test_LSTM import main_loop
+from hist_data_analysis.timeseries import main_loop
 from hist_data_analysis.correlate import correlate
