+++ hist_data_analysis/timeseries/__init__.py
@@ -1,3 +1,3 @@
 from hist_data_analysis.timeseries.data_loader import load_df, TimeSeriesDataset
-from hist_data_analysis.timeseries.train_and_test_mTAN import main
+from hist_data_analysis.timeseries.train_and_test_mTAN import main_loop
 from hist_data_analysis.timeseries.model import MtanGruRegr
